FINALLY fix the bug with the ampersand in mpris
So, what I failed to realize is that AwesomeWM uses *HTML* to format
elements, which includes escape codes! So when I tried to set the bar
title to "King Gizzard & The Lizard Wizard", it'd interpret the
ampersand as the beginning of an escape sequence, that'd never end...
Basically, my bar is vulnerable to HTML injection. But since I doubt
artists will be trying XSS through spotify song titles, I'm just
replacing & with &amp;. Should be enough for now. If XSS ever becomes an
actual thing, I'll fix it properly. For now though, this is good
enough™.